Application Id update. Crashlytics added
The app was removed from Google Play for sexually explicit material.
Since they didn't say exactly what was wrong, I'm going to assume it was
the fact that one of the screen shots had the word "Porn" in it (even
though it was "Earth Porn"). So the application id needs to be changed
to be re uploaded again.

- Crashlytics has been added to get better crash reports and usage data.

- NSFW Thumbnails are now able to be blocked if desirable.